Remove call_abi from process
